moving the sign in to it's own component, adding styling
